jsReplace, interceptor구현중

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -6,7 +6,7 @@
 	<section class="mt-8 text-xl">
 		<div class="container mx-auto px-3">
 			<div class="table-box-type-1">
-				<table border="1">
+				<table>
 					<colgroup>
 						<col width="200" />
 					</colgroup>
@@ -38,12 +38,13 @@
 					</tbody>
 				</table>
 			</div>
+			
 			<div class="btns">
-				<button type="button" onclick="history.back();">뒤로가기</button>
-					<c:if test="${article.actorCanChangeData }">
-						<a class="btn-text-link" href="modify?id=${article.id }">수정</a>
-						<a class="btn-text-link" href="doDelete?id=${article.id }">삭제</a>
-					</c:if>
+				<button class="btn-text-link" type="button" onclick="history.back();">뒤로가기</button>
+				<c:if test="${article.actorCanChangeData }">
+					<a class="btn-text-link" href="modify?id=${article.id }">수정</a>
+					<a class="btn-text-link" href="doDelete?id=${article.id }" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+				</c:if>
 			</div>
 		</div>
 	</section>
